requirements.txt
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-dramatiq==0.10.0
redis==4.3.1
requests==2.27.1
django-environ==0.8.1
